update
update

+++ aws_glue/source_etl_old/glue_script_catalog_repartition.py
@@ -9,7 +9,7 @@ from awsglue.job import Job
 import boto3
 
 # Glue/Spark context set up
-args = getResolvedOptions(sys.argv, ['JOB_NAME', 'glue_db', 's3_bucket', 'svr', 'db', 'sch', 'tbl', 's3_bucket_etc', 'api'])
+args = getResolvedOptions(sys.argv, ['JOB_NAME', 'glue_db', 's3_bucket', 'svr', 'db', 'sch', 'tbl', 's3_bucket_etc', 'partition_by', 'job-bookmark-option'])
 sc = SparkContext()
 glueContext = GlueContext(sc)
 spark = glueContext.spark_session
@@ -24,7 +24,8 @@ db = args['db']
 sch = args['sch']
 tbl = args['tbl']
 s3_bucket_etc = args['s3_bucket_etc']
-api = args['api']
+partition_by = args['partition_by']
+bookmark = args['job-bookmark-option']
 table_name = tbl
 s3_path ='s3://'+s3_bucket+'/'+s3_bucket_etc+'/'+sch+'/'+tbl
 prefix = s3_bucket_etc+'/'+sch+'/'+tbl
@@ -41,24 +42,22 @@ for column in tableColumns:
     mapping.append((column['Name'], column['Type'], column['Name'], column['Type']))
 
 # Delete prior runs s3 file
-s3 = boto3.resource('s3', region_name='us-west-2', verify=False)
-bucket = s3.Bucket(s3_bucket)
-objs = bucket.objects.filter(Prefix=prefix)
-for obj in objs:
-    print(obj.key)
-    obj.delete()
+if bookmark == 'job-bookmark-disable':
+    s3 = boto3.resource('s3', region_name='us-west-2', verify=False)
+    bucket = s3.Bucket(s3_bucket)
+    objs = bucket.objects.filter(Prefix=prefix)
+    for obj in objs:
+        print(obj.key)
+        obj.delete()
 
 # Run glue job
 ctx = tbl
 job.init(args['JOB_NAME'], args)
-datasource0 = glueContext.create_dynamic_frame.from_catalog(database=glue_db, table_name=table_name,
-                                                            transformation_ctx="datasource0")
-applymapping1 = ApplyMapping.apply(frame=datasource0, mappings=mapping, transformation_ctx="applymapping1")
-resolvechoice2 = ResolveChoice.apply(frame=applymapping1, choice="make_struct", transformation_ctx="resolvechoice2")
-dropnullfields3 = DropNullFields.apply(frame=resolvechoice2, transformation_ctx="dropnullfields3")
-datasink4 = glueContext.write_dynamic_frame.from_options(frame=dropnullfields3, connection_type="s3",
-                                                         connection_options={"path": s3_path, "partitionKeys": [api]},
-                                                         format="parquet", transformation_ctx="datasink4")
+datasource0 = glueContext.create_dynamic_frame.from_catalog(database=glue_db, table_name=table_name, transformation_ctx="datasource0"+ctx)
+applymapping1 = ApplyMapping.apply(frame=datasource0, mappings=mapping, transformation_ctx="applymapping1"+ctx)
+resolvechoice2 = ResolveChoice.apply(frame=applymapping1, choice="make_struct", transformation_ctx="resolvechoice2"+ctx)
+dropnullfields3 = DropNullFields.apply(frame=resolvechoice2, transformation_ctx="dropnullfields3"+ctx)
+datasink4 = glueContext.write_dynamic_frame.from_options(frame=dropnullfields3, connection_type="s3", connection_options={"path": s3_path, "partitionKeys": [partition_by]}, format="parquet", transformation_ctx="datasink4"+ctx)
 job.commit()
